+++ src/agenda/style.js
@@ -2,11 +2,13 @@ import {StyleSheet} from 'react-native';
 import * as defaultStyle from '../style';
 import platformStyles from './platform-style';
 
+
 const STYLESHEET_ID = 'stylesheet.agenda.main';
 
 export default function styleConstructor(theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
-  const { knob, weekdays } = platformStyles(appStyle);
+  const {knob, weekdays} = platformStyles(appStyle);
+  
   return StyleSheet.create({
     knob,
     weekdays,
@@ -15,12 +17,7 @@ export default function styleConstructor(theme = {}) {
       justifyContent: 'flex-end',
       position:'absolute',
       height:'100%',
-      width:'100%',
-    },
-    calendar: { // not in use
-      flex: 1,
-      borderBottomWidth: 1,
-      borderColor: appStyle.separatorColor
+      width:'100%'
     },
     knobContainer: {
       flex: 1,
@@ -35,8 +32,10 @@ export default function styleConstructor(theme = {}) {
     weekday: {
       width: 32,
       textAlign: 'center',
-      fontSize: 13,
       color: appStyle.textSectionTitleColor,
+      fontSize: appStyle.textDayHeaderFontSize,
+      fontFamily: appStyle.textDayHeaderFontFamily,
+      fontWeight: appStyle.textDayHeaderFontWeight
     },
     reservations: {
       flex: 1,