+++ src/calendar/day/multi-dot/index.js
@@ -83,7 +83,9 @@ class Day extends Component {
         testID={this.props.testID}
         style={containerStyle}
         onPress={this.onDayPress}
-        onLongPress={this.onDayLongPress}>
+        onLongPress={this.onDayLongPress}
+        disabled={marking.disableTouchEvent}
+      >
         <Text allowFontScaling={false} style={textStyle}>{String(this.props.children)}</Text>
         <View style={{flexDirection: 'row'}}>{dot}</View>
       </TouchableOpacity>