+++ src/expandableCalendar/week.js
@@ -16,7 +16,6 @@ import SingleDay from '../calendar/day/custom';
 import Calendar from '../calendar';
 
 
-const commons = require('./commons');
 const EmptyArray = [];
 
 class Week extends Component {
@@ -84,7 +83,6 @@ class Week extends Component {
 
   getDateMarking(day) {
     const {markedDates} = this.props;
-
     if (!markedDates) {
       return false;
     }
@@ -162,7 +160,7 @@ class Week extends Component {
 
     return (
       <View style={this.style.container}>
-        <View style={[this.style.week, this.props.style, {width: commons.screenWidth}]}>{week}</View>
+        <View style={[this.style.week, this.props.style]}>{week}</View>
       </View>
     );
   }